refactor: separate friends reducer into friends and users reducers
