+++ client/store/friends.js
@@ -4,37 +4,30 @@ import axios from 'axios'
  * ACTION TYPES
  */
 
-const GET_USERS = 'GET_USERS'
 const GET_FRIENDS = 'GET_FRIENDS'
 const ADD_FRIEND = 'ADD_FRIEND'
 
 /**
  * INITIAL STATE
  */
-const defaultUsers = {
-  users: [],
-  friends: []
-}
+const defaultFriends = []
 
 /**
  * ACTION CREATORS
  */
-const getUsers = users => ({type: GET_USERS, users})
-const getFriends = friends => ({type: GET_FRIENDS, friends})
-const addNewFriend = friend => ({type: ADD_FRIEND, friend})
+
+const getFriends = friends => ({
+  type: GET_FRIENDS,
+  friends
+})
+const addNewFriend = friend => ({
+  type: ADD_FRIEND,
+  friend
+})
 
 /**
  * THUNK CREATORS
  */
-
-export const fetchUsersFromDB = () => async dispatch => {
-  try {
-    const {data: users} = await axios.get(`/api/users/`)
-    dispatch(getUsers(users))
-  } catch (err) {
-    console.error(err)
-  }
-}
 
 export const fetchFriends = () => async dispatch => {
   try {
@@ -57,14 +50,12 @@ export const addFriend = id => async dispatch => {
 /**
  * REDUCER
  */
-export default function(state = defaultUsers, action) {
+export default function(state = defaultFriends, action) {
   switch (action.type) {
-    case GET_USERS:
-      return {...state, users: action.users}
     case GET_FRIENDS:
-      return {...state, friends: action.friends}
+      return action.friends
     case ADD_FRIEND:
-      return {...state, friends: [...state.friends, action.friend]}
+      return [...state, action.friend]
     default:
       return state
   }
